Move function out from API/common/utils.py
Move not common functions to API/builder/ and API/testrunner/ to separate functionality correctly.

JSRemoteTest-DCO-1.0-Signed-off-by: Bela Toth tbela@inf.u-szeged.hu
