+++ API/testrunner/testrunner.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 from API.common import paths, reporter, utils
+from API.testrunner import utils as testrunnerUtils
 from API.testrunner import devices
 from API.testrunner.skiplist import Skiplist
 
@@ -26,7 +27,7 @@ def read_testsets(env):
     # Since JerryScript doesn't have testset descriptor file,
     # simply read the file contents from the test folder.
     if application['name'] == 'jerryscript':
-        testsets = utils.read_test_files(env)
+        testsets = testrunnerUtils.read_test_files(env)
     else:
         testsets = utils.read_json_file(application['paths']['testfiles'])
 
@@ -72,7 +73,7 @@ class TestRunner(object):
                 result_dir =  utils.join(paths.RESULT_PATH, '%s/%s/' % (app_name, device))
                 result_path = utils.join(result_dir, result_name)
 
-            self.coverage_info = utils.parse_coverage_info(self.env, result_path)
+            self.coverage_info = testrunnerUtils.parse_coverage_info(self.env, result_path)
 
             reporter.report_coverage(self.coverage_info)
 
@@ -161,4 +162,4 @@ class TestRunner(object):
         utils.write_json_file(filename + '.json', test_info)
 
         # Publish the results if necessary.
-        utils.upload_data_to_firebase(self.env, test_info)
+        testrunnerUtils.upload_data_to_firebase(self.env, test_info)
